feat: Migrate api/capture to app dir
